pyproject.toml
[project]
name = "test_publish"
version = "0.1.1"
description = "Test publish"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fast-bitrix24",
    "python-dotenv",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[project.urls]
Homepage = "https://github.com/darkClaw921/test_publish"
Documentation = "https://github.com/darkClaw921/test_publish"
Repository = "https://github.com/darkClaw921/test_publish.git"
"Bug Tracker" = "https://github.com/darkClaw921/test_publish/issues"
Changelog = "https://github.com/darkClaw921/test_publish/blob/master/CHANGELOG.md"



[dependency-groups]
dev = [
    "loguru>=0.7.3",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
]
